kosten functie van unsigned long naar int voor meer reken gemak

+++ include/compress.h
@@ -489,8 +489,8 @@ typedef struct packstruct_t          /* Bij aanpassing van deze struct ook ENCOD
   long (*buf_read_crc)(long count, void *buf, void* brc_propagator); /* read routine for reading the original file */
   void* brc_propagator;
   gup_result (*close_packed_stream)(struct packstruct_t *com); /* output the last bits in bitbuf */
-  unsigned long (*cost_ptrlen)(match_t match, ptr_t ptr, index_t pos, ptr_t *ptr_hist); /* costs of pointer len combi */
-  unsigned long (*cost_lit)(match_t kar); /* costs of a literal */
+  int (*cost_ptrlen)(match_t match, ptr_t ptr, index_t pos, ptr_t *ptr_hist); /* costs of pointer len combi */
+  int (*cost_lit)(match_t kar); /* costs of a literal */
   uint8 *inmem_output; /* buffer, origsize*2+8 bytes in size, after 8 bytes it shares space with the inmem_input buffer */
   uint8 *inmem_output_cur; /* current position in inmem_ouput buffer */
   uint16 *inmem_input; /* buffer for origsize 16 bit words */
